# apps/api/app/api/routes/recoveries.py
"""Recovery candidates: turn a validated treaty + a loss event into a reviewable,
deterministically-calculated recovery. No AI in this module (ADR-0010)."""

from __future__ import annotations

from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Depends, Query, status
from fastapi.responses import HTMLResponse

from app.api.dependencies.context import (
    AppSettings,
    AuthedContext,
    DbSession,
    InvestigateEnqueuer,
    NoticeEnqueuer,
    get_investigate_enqueuer,
    get_notice_enqueuer,
)
from app.api.schemas.recoveries import (
    AgentRunToolCalls,
    CalcStepOut,
    CalculationAllocationOut,
    CreateRecoveryCandidateRequest,
    DraftNoticeRequest,
    GeneratePacketResponse,
    InvestigationCitationOut,
    InvestigationFindingOut,
    NoticeReviewRequest,
    PacketCitationOut,
    PacketContentOut,
    PacketReviewRequest,
    PacketSectionOut,
    PacketStatementOut,
    RecoveryCalculationOut,
    RecoveryCandidateDetail,
    RecoveryCandidateList,
    RecoveryCandidateOut,
    RecoveryInvestigationOut,
    RecoveryNoticeList,
    RecoveryNoticeOut,
    RecoveryPacketDetail,
    RecoveryPacketVersionOut,
    RecoveryPacketVersionSummary,
    RecoveryReviewOut,
    ReviewRecoveryCandidateRequest,
    ToolCallOut,
)
from app.db.models.extraction import Review, ToolCall
from app.db.models.recoveries import (
    RecoveryCalculation,
    RecoveryCandidate,
    RecoveryInvestigation,
    RecoveryNotice,
    RecoveryPacket,
    RecoveryPacketVersion,
)
from app.domain.recoveries import RecoveryCandidateStatus
from app.services.errors import ConflictError
from app.services.investigation import InvestigationService
from app.services.notice import NoticeReview, NoticeService
from app.services.packet import PacketReview, RecoveryPacketService
from app.services.recoveries import RecoveryCandidateService

router = APIRouter(prefix="/recovery-candidates", tags=["recovery-candidates"])
packets_router = APIRouter(prefix="/recovery-packets", tags=["recovery-packets"])
notices_router = APIRouter(prefix="/recovery-notices", tags=["recovery-notices"])


def _candidate_out(candidate: RecoveryCandidate) -> RecoveryCandidateOut:
    return RecoveryCandidateOut(
        id=candidate.id,
        status=candidate.status,
        treaty_id=candidate.treaty_id,
        treaty_version_id=candidate.treaty_version_id,
        treaty_layer_id=candidate.treaty_layer_id,
        loss_event_id=candidate.loss_event_id,
        currency=candidate.currency,
        gross_event_incurred=candidate.gross_event_incurred,
        currency_mismatch=candidate.currency_mismatch,
        current_calculation_id=candidate.current_calculation_id,
        created_at=candidate.created_at,
        reviewed_at=candidate.reviewed_at,
    )


def _calculation_out(calc: RecoveryCalculation) -> RecoveryCalculationOut:
    return RecoveryCalculationOut(
        id=calc.id,
        engine_version=calc.engine_version,
        currency=calc.currency,
        gross_loss=calc.gross_loss,
        attachment=calc.attachment,
        amount_above_attachment=calc.amount_above_attachment,
        layer_limit=calc.layer_limit,
        layer_recovery=calc.layer_recovery,
        cedent_retention=calc.cedent_retention,
        total_ceded=calc.total_ceded,
        input_hash=calc.input_hash,
        trace=[
            CalcStepOut(label=s["label"], expression=s["expression"], result=s["result"])
            for s in calc.trace
        ],
        allocations=[
            CalculationAllocationOut(
                reinsurer_id=a.reinsurer_id,
                reinsurer_name=a.reinsurer.name,
                participation_share=a.participation_share,
                allocated_recovery=a.allocated_recovery,
            )
            for a in calc.allocations
        ],
        created_at=calc.created_at,
    )


def _review_out(review: Review) -> RecoveryReviewOut:
    return RecoveryReviewOut(
        decision=review.decision, reason=review.reason, created_at=review.created_at
    )


def _investigation_out(inv: RecoveryInvestigation) -> RecoveryInvestigationOut:
    return RecoveryInvestigationOut(
        id=inv.id,
        status=inv.status,
        agent_run_id=inv.agent_run_id,
        summary=inv.summary,
        applicability_assessment=inv.applicability_assessment,
        confidence=float(inv.confidence) if inv.confidence is not None else None,
        out_of_scope=inv.out_of_scope,
        suspected_prompt_injection=inv.suspected_prompt_injection,
        unresolved_questions=list(inv.unresolved_questions),
        superseded=inv.superseded_at is not None,
        created_at=inv.created_at,
        findings=[
            InvestigationFindingOut(
                ordinal=f.ordinal,
                kind=f.kind,
                text=f.text,
                confidence=float(f.confidence) if f.confidence is not None else None,
                citation=(
                    InvestigationCitationOut(
                        document_id=f.citation.document_id,
                        page_number=f.citation.page_number,
                        section=f.citation.section,
                        quoted_text=f.citation.quoted_text,
                    )
                    if f.citation is not None
                    else None
                ),
            )
            for f in inv.findings
        ],
    )


def _tool_call_out(call: ToolCall) -> ToolCallOut:
    return ToolCallOut(
        ordinal=call.ordinal,
        tool_name=call.tool_name,
        arguments=dict(call.arguments),
        result_summary=dict(call.result_summary),
        status=call.status.value,
    )


@router.post(
    "",
    response_model=RecoveryCandidateOut,
    status_code=status.HTTP_201_CREATED,
    summary="Create a recovery candidate (validated treaty + loss event → deterministic calc)",
    operation_id="createRecoveryCandidate",
)
async def create_recovery_candidate(
    payload: CreateRecoveryCandidateRequest, context: AuthedContext, session: DbSession
) -> RecoveryCandidateOut:
    candidate = await RecoveryCandidateService(session).create(
        context, treaty_id=payload.treaty_id, loss_event_id=payload.loss_event_id
    )
    return _candidate_out(candidate)


@router.get("", response_model=RecoveryCandidateList, operation_id="listRecoveryCandidates")
async def list_recovery_candidates(
    context: AuthedContext,
    session: DbSession,
    status_filter: Annotated[RecoveryCandidateStatus | None, Query(alias="status")] = None,
) -> RecoveryCandidateList:
    candidates = await RecoveryCandidateService(session).list_candidates(
        context, status=status_filter
    )
    return RecoveryCandidateList(candidates=[_candidate_out(c) for c in candidates])


@router.get(
    "/{candidate_id}", response_model=RecoveryCandidateDetail, operation_id="getRecoveryCandidate"
)
async def get_recovery_candidate(
    candidate_id: UUID, context: AuthedContext, session: DbSession, settings: AppSettings
) -> RecoveryCandidateDetail:
    service = RecoveryCandidateService(session)
    candidate = await service.get_candidate(context, candidate_id)
    reviews = await service.candidate_reviews(context, candidate_id)
    current = service.current_calculation(candidate)
    investigations = await InvestigationService(session, settings).list_for_candidate(
        context, candidate_id
    )
    return RecoveryCandidateDetail(
        candidate=_candidate_out(candidate),
        current_calculation=_calculation_out(current) if current else None,
        calculations=[
            _calculation_out(c)
            for c in sorted(candidate.calculations, key=lambda x: x.created_at, reverse=True)
        ],
        reviews=[_review_out(r) for r in reviews],
        investigations=[_investigation_out(i) for i in investigations],
    )


@router.post(
    "/{candidate_id}/recalculate",
    response_model=RecoveryCandidateOut,
    summary="Re-run the engine; a new immutable calculation is stored only if inputs changed",
    operation_id="recalculateRecoveryCandidate",
)
async def recalculate_recovery_candidate(
    candidate_id: UUID, context: AuthedContext, session: DbSession
) -> RecoveryCandidateOut:
    candidate = await RecoveryCandidateService(session).recalculate(context, candidate_id)
    return _candidate_out(candidate)


@router.post(
    "/{candidate_id}/review",
    response_model=RecoveryCandidateOut,
    summary="Human decision: confirm | reject | request_info",
    operation_id="reviewRecoveryCandidate",
)
async def review_recovery_candidate(
    candidate_id: UUID,
    payload: ReviewRecoveryCandidateRequest,
    context: AuthedContext,
    session: DbSession,
) -> RecoveryCandidateOut:
    candidate = await RecoveryCandidateService(session).review(
        context, candidate_id, decision=payload.decision, reason=payload.reason
    )
    return _candidate_out(candidate)


@router.post(
    "/{candidate_id}/investigate",
    response_model=RecoveryCandidateOut,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Queue the Recovery Investigator (bounded, read-only AI — never computes the recovery)",
    operation_id="investigateRecoveryCandidate",
)
async def investigate_recovery_candidate(
    candidate_id: UUID,
    context: AuthedContext,
    session: DbSession,
    settings: AppSettings,
    enqueue: Annotated[InvestigateEnqueuer, Depends(get_investigate_enqueuer)],
) -> RecoveryCandidateOut:
    if not settings.ai_enabled:
        raise ConflictError("AI is disabled in this environment")
    # Surfaces 404 / 409 before the job is queued.
    candidate = await RecoveryCandidateService(session).get_candidate(context, candidate_id)
    if candidate.current_calculation_id is None:
        raise ConflictError("the candidate has no calculation to investigate")
    await enqueue(context.organization.id, candidate.id, context.user.id)
    return _candidate_out(candidate)


@router.get(
    "/{candidate_id}/agent-runs/{agent_run_id}/tool-calls",
    response_model=AgentRunToolCalls,
    operation_id="getRecoveryAgentToolCalls",
)
async def get_recovery_agent_tool_calls(
    candidate_id: UUID,
    agent_run_id: UUID,
    context: AuthedContext,
    session: DbSession,
    settings: AppSettings,
) -> AgentRunToolCalls:
    calls = await InvestigationService(session, settings).tool_calls(context, agent_run_id)
    return AgentRunToolCalls(
        agent_run_id=agent_run_id, tool_calls=[_tool_call_out(c) for c in calls]
    )


# --- recovery packet ----------------------------------------------


def _packet_content_out(content: dict) -> PacketContentOut:
    return PacketContentOut(
        title=content["title"],
        subtitle=content["subtitle"],
        generated_at=content["generated_at"],
        engine_version=content["engine_version"],
        sections=[
            PacketSectionOut(
                key=section["key"],
                title=section["title"],
                statements=[
                    PacketStatementOut(
                        key=s["key"],
                        statement_class=s["statement_class"],
                        text=s["text"],
                        citation=(PacketCitationOut(**s["citation"]) if s["citation"] else None),
                        detail=s.get("detail", {}),
                        edited_by_human=s.get("edited_by_human", False),
                    )
                    for s in section["statements"]
                ],
            )
            for section in content["sections"]
        ],
    )


def _packet_version_out(version: RecoveryPacketVersion) -> RecoveryPacketVersionOut:
    return RecoveryPacketVersionOut(
        id=version.id,
        version_no=version.version_no,
        status=version.status,
        calculation_id=version.calculation_id,
        investigation_id=version.investigation_id,
        review_note=version.review_note,
        approved_at=version.approved_at,
        superseded=version.superseded_at is not None,
        created_at=version.created_at,
        content=_packet_content_out(version.content),
    )


def _packet_detail(
    packet: RecoveryPacket, current: RecoveryPacketVersion | None
) -> RecoveryPacketDetail:
    return RecoveryPacketDetail(
        packet_id=packet.id,
        recovery_candidate_id=packet.recovery_candidate_id,
        human_overrides=dict(packet.human_overrides),
        current_version=_packet_version_out(current) if current else None,
        versions=[
            RecoveryPacketVersionSummary(
                id=v.id,
                version_no=v.version_no,
                status=v.status,
                superseded=v.superseded_at is not None,
                created_at=v.created_at,
            )
            for v in sorted(packet.versions, key=lambda x: x.version_no, reverse=True)
        ],
    )


@router.post(
    "/{candidate_id}/packet",
    response_model=GeneratePacketResponse,
    status_code=status.HTTP_201_CREATED,
    summary="Assemble a recovery packet (a new immutable version each time)",
    operation_id="generateRecoveryPacket",
)
async def generate_recovery_packet(
    candidate_id: UUID, context: AuthedContext, session: DbSession
) -> GeneratePacketResponse:
    version = await RecoveryPacketService(session).generate(context, candidate_id)
    return GeneratePacketResponse(
        packet_id=version.recovery_packet_id, version=_packet_version_out(version)
    )


@router.get(
    "/{candidate_id}/packet",
    response_model=RecoveryPacketDetail,
    operation_id="getRecoveryPacket",
)
async def get_recovery_packet(
    candidate_id: UUID, context: AuthedContext, session: DbSession
) -> RecoveryPacketDetail:
    service = RecoveryPacketService(session)
    packet = await service.get_for_candidate(context, candidate_id)
    return _packet_detail(packet, service.current_version(packet))


@packets_router.post(
    "/{packet_id}/versions/{version_id}/review",
    response_model=RecoveryPacketVersionOut,
    summary="Human decision on a packet version: confirm | reject | request_info | edit",
    operation_id="reviewRecoveryPacketVersion",
)
async def review_recovery_packet_version(
    packet_id: UUID,
    version_id: UUID,
    payload: PacketReviewRequest,
    context: AuthedContext,
    session: DbSession,
) -> RecoveryPacketVersionOut:
    version = await RecoveryPacketService(session).review_version(
        context,
        packet_id,
        version_id,
        PacketReview(
            decision=payload.decision,
            reason=payload.reason,
            statement_key=payload.statement_key,
            value=payload.value,
        ),
    )
    return _packet_version_out(version)


@packets_router.get(
    "/{packet_id}/versions/{version_id}/html",
    response_class=HTMLResponse,
    summary="The rendered packet HTML (for printing / archiving)",
    operation_id="getRecoveryPacketHtml",
)
async def get_recovery_packet_html(
    packet_id: UUID, version_id: UUID, context: AuthedContext, session: DbSession
) -> HTMLResponse:
    version = await RecoveryPacketService(session).version_html(context, version_id)
    return HTMLResponse(
        content=version.rendered_html or "<!doctype html><title>Empty packet</title>",
        headers={"Cache-Control": "private, no-store"},
    )


# --- recovery notice ----------------------------------------------


def _notice_out(notice: RecoveryNotice) -> RecoveryNoticeOut:
    return RecoveryNoticeOut(
        id=notice.id,
        recovery_candidate_id=notice.recovery_candidate_id,
        kind=notice.kind,
        status=notice.status,
        recipient=dict(notice.recipient),
        subject=notice.subject,
        body_markdown=notice.body_markdown,
        key_figures=dict(notice.key_figures),
        caveats=list(notice.caveats),
        used_only_provided_facts=notice.used_only_provided_facts,
        notes_for_reviewer=notice.notes_for_reviewer,
        context=dict(notice.context),
        agent_run_id=notice.agent_run_id,
        recovery_packet_version_id=notice.recovery_packet_version_id,
        review_note=notice.review_note,
        approved_at=notice.approved_at,
        superseded=notice.superseded_at is not None,
        created_at=notice.created_at,
    )


@router.post(
    "/{candidate_id}/notices",
    response_model=RecoveryCandidateOut,
    status_code=status.HTTP_202_ACCEPTED,
    summary="Queue the notice drafter (whitelist of approved facts in; draft out; never sent)",
    operation_id="draftRecoveryNotice",
)
async def draft_recovery_notice(
    candidate_id: UUID,
    payload: DraftNoticeRequest,
    context: AuthedContext,
    session: DbSession,
    settings: AppSettings,
    enqueue: Annotated[NoticeEnqueuer, Depends(get_notice_enqueuer)],
) -> RecoveryCandidateOut:
    if not settings.ai_enabled:
        raise ConflictError("AI is disabled in this environment")
    candidate = await RecoveryCandidateService(session).get_candidate(context, candidate_id)
    if candidate.status.value not in ("confirmed", "notice_drafted"):
        raise ConflictError("confirm the recovery candidate before drafting a notice")
    await enqueue(
        context.organization.id,
        candidate.id,
        kind=payload.kind.value,
        recipient=payload.recipient.model_dump(),
        actor_id=context.user.id,
    )
    return _candidate_out(candidate)


@router.get(
    "/{candidate_id}/notices",
    response_model=RecoveryNoticeList,
    operation_id="listRecoveryNotices",
)
async def list_recovery_notices(
    candidate_id: UUID, context: AuthedContext, session: DbSession, settings: AppSettings
) -> RecoveryNoticeList:
    notices = await NoticeService(session, settings).list_for_candidate(context, candidate_id)
    return RecoveryNoticeList(notices=[_notice_out(n) for n in notices])


@notices_router.get(
    "/{notice_id}", response_model=RecoveryNoticeOut, operation_id="getRecoveryNotice"
)
async def get_recovery_notice(
    notice_id: UUID, context: AuthedContext, session: DbSession, settings: AppSettings
) -> RecoveryNoticeOut:
    notice = await NoticeService(session, settings).get_notice(context, notice_id)
    return _notice_out(notice)


@notices_router.post(
    "/{notice_id}/review",
    response_model=RecoveryNoticeOut,
    summary="Human decision on a draft notice: confirm | reject | request_info | edit",
    operation_id="reviewRecoveryNotice",
)
async def review_recovery_notice(
    notice_id: UUID,
    payload: NoticeReviewRequest,
    context: AuthedContext,
    session: DbSession,
    settings: AppSettings,
) -> RecoveryNoticeOut:
    notice = await NoticeService(session, settings).review(
        context,
        notice_id,
        NoticeReview(
            decision=payload.decision,
            subject=payload.subject,
            body_markdown=payload.body_markdown,
            recipient=payload.recipient.model_dump() if payload.recipient else None,
            reason=payload.reason,
        ),
    )
    return _notice_out(notice)