# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRecoveryCandidateRequest": {
        "additionalProperties": false,
        "properties": {
          "loss_event_id": {
            "format": "uuid",
            "title": "Loss Event Id",
            "type": "string"
          },
          "treaty_id": {
            "format": "uuid",
            "title": "Treaty Id",
            "type": "string"
          }
        },
        "required": [
          "treaty_id",
          "loss_event_id"
        ],
        "title": "CreateRecoveryCandidateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecoveryCandidateList": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RecoveryCandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "required": [
          "candidates"
        ],
        "title": "RecoveryCandidateList",
        "type": "object"
      },
...
      "RecoveryCandidateStatus": {
        "description": "docs/PRODUCT.md \u2014 Recovery Candidate lifecycle.\n\nNEEDS_REVIEW \u2500\u25b6 IN_REVIEW \u2500\u25b6 CONFIRMED \u2500\u25b6 NOTICE_DRAFTED\n                      \u2514\u2500\u2500\u25b6 REJECTED\n(inputs change \u2192 recalculation \u2192 new immutable RecoveryCalculation;\n a CONFIRMED candidate reverts to NEEDS_REVIEW)",
        "enum": [
          "draft",
          "needs_review",
          "in_review",
          "confirmed",
          "rejected",
          "notice_drafted"
        ],
        "title": "RecoveryCandidateStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Reinsurance intelligence from contract to recovery.",
    "title": "Cedeon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/recovery-candidates": {
      "get": {
        "operationId": "listRecoveryCandidates",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecoveryCandidateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recovery Candidates",
        "tags": [
          "recovery-candidates"
        ]
      },
      "post": {
        "operationId": "createRecoveryCandidate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecoveryCandidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a recovery candidate (validated treaty + loss event \u2192 deterministic calc)",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}": {
      "get": {
        "operationId": "getRecoveryCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Candidate",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/agent-runs/{agent_run_id}/tool-calls": {
      "get": {
        "operationId": "getRecoveryAgentToolCalls",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "agent_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunToolCalls"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Agent Tool Calls",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/investigate": {
      "post": {
        "operationId": "investigateRecoveryCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue the Recovery Investigator (bounded, read-only AI \u2014 never computes the recovery)",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/notices": {
      "get": {
        "operationId": "listRecoveryNotices",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryNoticeList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recovery Notices",
        "tags": [
          "recovery-candidates"
        ]
      },
      "post": {
        "operationId": "draftRecoveryNotice",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftNoticeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue the notice drafter (whitelist of approved facts in; draft out; never sent)",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/packet": {
      "get": {
        "operationId": "getRecoveryPacket",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPacketDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Packet",
        "tags": [
          "recovery-candidates"
        ]
      },
      "post": {
        "operationId": "generateRecoveryPacket",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePacketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assemble a recovery packet (a new immutable version each time)",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/recalculate": {
      "post": {
        "operationId": "recalculateRecoveryCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Re-run the engine; a new immutable calculation is stored only if inputs changed",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-candidates/{candidate_id}/review": {
      "post": {
        "operationId": "reviewRecoveryCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRecoveryCandidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Human decision: confirm | reject | request_info",
        "tags": [
          "recovery-candidates"
        ]
      }
    },
    "/recovery-notices/{notice_id}": {
      "get": {
        "operationId": "getRecoveryNotice",
        "parameters": [
          {
            "in": "path",
            "name": "notice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryNoticeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Notice",
        "tags": [
          "recovery-notices"
        ]
      }
    },
...
  }
}